requirements.txt
requests===2.31.0
discord.py===2.3.2
mysql-connector-python===8.4.0
python-dotenv===1.0.1